lettuce: sensors: ensure first sample is always reported
Fix a race condition between activate/setEnable and poll. The 'enabled'
status is stored in the SensorContext* in NativeSensorManager::activate
after the driver has been enabled. This sometimes results in a new
sensor sample arriving in poll( because enable is not yet set.

For continuous sensors, this just results in a dropped first sample and
is not a severe issue. However on-change sensors cannot recover from
this state as another sample might not be received for a while.

All operations are now run on the poll thread and the wake pipe is used
as a command pipe. HAL interface operations are now sent as command
structs through the pipe and processed in the poll thread.

Make the proximity sensor re-report the previous sample when it's first
enabled as Android expects the current state to be reported.

This change fixes a CTS v5.1r3 failure in SensorTest#testBatchAndFlush
(LETTUCE-779)

Change-Id: Id6526a16003c3d57ce52a179587f25a5069c913e
Signed-off-by: Ad!thya R <adithya.r02@outlook.com>
Signed-off-by: Ad!thya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: ghostrider-reborn <adithya.r02@outlook.com>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>