lettuce: sensors: fix prox sensor reporting 'near' on 1st enable
The driver always reports a 'far' abs value when it's first enabled,
the HAL misses this event if does not result in an abs value change.

With the addition of Kernel based timestamps, a new timestamp value
would be generated triggering a sensor event being reported to
Android. The mPendingEvent.distance field would be out of sync with
the abs input dev's distance value.

Handle this case by querying the abs value from the input device when
the sensor is enabled.

LETTUCE-779

Change-Id: I53bfa0d10f92efff8ec92abc7fc8df6a589098a7
Signed-off-by: Ad!thya R <adithya.r02@outlook.com>
Signed-off-by: Ad!thya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: ghostrider-reborn <adithya.r02@outlook.com>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>